Made notes fall to the rhythm of the music

+++ FNF/BeatMap.swift
@@ -8,20 +8,20 @@ import UIKit
 
 struct Note {
     // Properties of any given note
-    let direction: CGFloat
-    let length: CGFloat
+    let direction: Int
+    let length: Int
     
     // Whether a note needs be held down or not
     let held: Bool
     // Value and length for hold note i.e Held length = half, HeldNum = 3 --> Hold note for 3 half notes
-    let heldLength: CGFloat
-    let heldNum: CGFloat
+    let heldLength: Int
+    let heldNum: Int
 }
 
 struct BeatMap {
     
     // Beats per minute -- Used to determine delay between notes
-    let bpm: CGFloat
+    let bpm: Double
     
     // Name to be displayed as song name
     let songName: String
@@ -30,7 +30,7 @@ struct BeatMap {
     let fileName: String
     
     // Time delay before first note of song
-    let startDelay: CGFloat
+    let startDelay: Double
     
     var notes: [Note]
 }
@@ -39,18 +39,18 @@ struct BeatMap {
 
 //Direction will tell us where to place note and what button it will need to interact with
 enum directions{
-    static let Left: CGFloat = 0
-    static let Right: CGFloat = 1
-    static let Up: CGFloat = 2
-    static let Down: CGFloat = 3
+    static let Left: Int = 0
+    static let Right: Int = 1
+    static let Up: Int = 2
+    static let Down: Int = 3
 }
 
 //Length will tell us how
 enum lengths{
-    static let Whole: CGFloat = 0
-    static let Half: CGFloat = 1
-    static let Quarter: CGFloat = 2
-    static let Eighth: CGFloat = 3
-    static let Sixteenth: CGFloat = 4
-    static let ThirtySecond: CGFloat = 5
+    static let Whole: Int = 0
+    static let Half: Int = 1
+    static let Quarter: Int = 2
+    static let Eighth: Int = 3
+    static let Sixteenth: Int = 4
+    static let ThirtySecond: Int = 5
 }
